+++ src/core/refactor/move_items.rs
@@ -1,15 +1,3 @@
-mod extension_integration;
-mod item_kind;
-mod move_options;
-mod types;
-mod whole-file_move;
-
-pub use extension_integration::*;
-pub use item_kind::*;
-pub use move_options::*;
-pub use types::*;
-pub use whole-file_move::*;
-
 //! Refactor move — extract items from one file and move them to another.
 //!
 //! Language-agnostic orchestration layer. All language-specific parsing
@@ -34,6 +22,75 @@ use crate::extension::{
 };
 use crate::{component, Result};
 
+/// Result of a move operation.
+#[derive(Debug, Clone, serde::Serialize)]
+pub struct MoveResult {
+    /// Items that were moved.
+    pub items_moved: Vec<MovedItem>,
+    /// The source file items were extracted from.
+    pub from_file: String,
+    /// The destination file items were moved to.
+    pub to_file: String,
+    /// Whether the destination file was created (vs. appended to).
+    pub file_created: bool,
+    /// Number of import references updated across the codebase.
+    pub imports_updated: usize,
+    /// Absolute paths of caller files whose imports were rewritten.
+    /// Used by decompose rollback to restore these files if the move is reverted.
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub caller_files_modified: Vec<PathBuf>,
+    /// Related tests that were moved alongside items.
+    pub tests_moved: Vec<MovedItem>,
+    /// Whether changes were written to disk.
+    pub applied: bool,
+    /// Warnings generated during the move.
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub warnings: Vec<String>,
+}
+
+/// Behavioral options for move operations.
+#[derive(Debug, Clone, Copy)]
+pub struct MoveOptions {
+    /// Whether related test functions should be moved alongside requested items.
+    pub move_related_tests: bool,
+}
+
+impl Default for MoveOptions {
+    fn default() -> Self {
+        Self {
+            move_related_tests: true,
+        }
+    }
+}
+
+/// A single item that was moved.
+#[derive(Debug, Clone, serde::Serialize)]
+pub struct MovedItem {
+    /// Name of the item (function, struct, etc.).
+    pub name: String,
+    /// What kind of item.
+    pub kind: ItemKind,
+    /// Line range in the source file (1-indexed, inclusive).
+    pub source_lines: (usize, usize),
+    /// Number of lines (including doc comments and attributes).
+    pub line_count: usize,
+}
+
+#[derive(Debug, Clone, serde::Serialize)]
+#[serde(rename_all = "snake_case")]
+pub enum ItemKind {
+    Function,
+    Struct,
+    Enum,
+    Const,
+    Static,
+    TypeAlias,
+    Impl,
+    Trait,
+    Test,
+    Unknown,
+}
+
 impl ItemKind {
     fn from_str(s: &str) -> Self {
         match s {
@@ -54,6 +111,24 @@ impl ItemKind {
 // ============================================================================
 // Extension Integration
 // ============================================================================
+
+/// Find a refactor-capable extension for a file based on its extension.
+fn find_refactor_extension(file_path: &str) -> Option<ExtensionManifest> {
+    let ext = Path::new(file_path).extension().and_then(|e| e.to_str())?;
+    extension::find_extension_for_file_ext(ext, "refactor")
+}
+
+/// Try parsing items using the core grammar engine (no extension script needed).
+fn core_parse_items(ext: &ExtensionManifest, content: &str) -> Option<Vec<ParsedItem>> {
+    let ext_path = ext.extension_path.as_deref()?;
+    let file_ext = ext.provided_file_extensions().first()?.clone();
+    let grammar = load_extension_grammar(Path::new(ext_path), &file_ext)?;
+    let items = grammar_items::parse_items(content, &grammar);
+    if items.is_empty() {
+        return None;
+    }
+    Some(items.into_iter().map(ParsedItem::from).collect())
+}
 
 /// Ask an extension to parse all top-level items in a source file.
 fn ext_parse_items(
@@ -172,6 +247,26 @@ pub fn ext_generate_module_index(
     });
     let result = extension::run_refactor_script(&ext, &cmd)?;
     result.get("content")?.as_str().map(|s| s.to_string())
+}
+
+/// A submodule entry for module index generation.
+#[derive(Debug, Clone)]
+pub struct ModuleIndexEntry {
+    /// Module name (e.g., "types", "unreleased").
+    pub name: String,
+    /// Public items that should be re-exported. Empty = glob re-export.
+    pub pub_items: Vec<String>,
+}
+
+/// A single import rewrite in a caller file.
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+pub struct ImportRewrite {
+    /// Line number (1-indexed) in the file.
+    pub line: usize,
+    /// Original line text.
+    pub original: String,
+    /// Replacement line text.
+    pub replacement: String,
 }
 
 // ============================================================================
@@ -659,9 +754,368 @@ pub fn move_items_with_options(
 // Whole-file move
 // ============================================================================
 
+/// Result of a whole-file move operation.
+#[derive(Debug, Clone, serde::Serialize)]
+pub struct MoveFileResult {
+    /// The source file that was moved.
+    pub from_file: String,
+    /// The destination file.
+    pub to_file: String,
+    /// Number of import references updated across the codebase.
+    pub imports_updated: usize,
+    /// Files whose imports were rewritten.
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub caller_files_modified: Vec<String>,
+    /// Whether changes were written to disk.
+    pub applied: bool,
+    /// Warnings generated during the move.
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub warnings: Vec<String>,
+    /// Whether mod declarations were updated.
+    pub mod_declarations_updated: bool,
+}
+
+/// Move an entire module file to a new location, rewriting all imports.
+///
+/// This is the `refactor move --file` operation:
+/// 1. Move the file to the new path
+/// 2. Remove `mod foo;` from the old parent's mod.rs
+/// 3. Add `mod foo;` to the new parent's mod.rs
+/// 4. Rewrite all `use crate::old::path::module` → `use crate::new::path::module` across the codebase
+pub fn move_file(from: &str, to: &str, root: &Path, write: bool) -> Result<MoveFileResult> {
+    let source_abs = root.join(from);
+    let dest_abs = root.join(to);
+    let warnings = Vec::new();
+
+    // Validate source exists
+    if !source_abs.exists() {
+        return Err(crate::Error::validation_invalid_argument(
+            "from",
+            format!("Source file does not exist: {}", from),
+            None,
+            None,
+        ));
+    }
+
+    // Validate destination doesn't exist
+    if dest_abs.exists() {
+        return Err(crate::Error::validation_invalid_argument(
+            "to",
+            format!("Destination already exists: {}", to),
+            None,
+            None,
+        ));
+    }
+
+    // Compute module paths for import rewriting
+    let source_module = module_path_from_file(from);
+    let dest_module = module_path_from_file(to);
+
+    // Derive module names for mod.rs updates
+    let source_stem = Path::new(from)
+        .file_stem()
+        .and_then(|s| s.to_str())
+        .unwrap_or("");
+    let dest_stem = Path::new(to)
+        .file_stem()
+        .and_then(|s| s.to_str())
+        .unwrap_or("");
+
+    // Find the extension for import rewriting
+    let ext = find_refactor_extension(from);
+
+    // ── Phase 1: Rewrite caller imports across the codebase ──────────
+    let mut imports_updated: usize = 0;
+    let mut caller_rewrites: Vec<(PathBuf, Vec<ImportRewrite>)> = Vec::new();
+
+    if let Some(ref ext) = ext {
+        if source_module != dest_module {
+            // For a whole-module move, we need to find all files that import
+            // anything from the source module. We use the module name as the
+            // search term — any file that mentions it might have imports to update.
+            let source_module_leaf = source_module.rsplit("::").next().unwrap_or(&source_module);
+
+            let all_files = codebase_scan::walk_files(
+                root,
+                &ScanConfig {
+                    extensions: ExtensionFilter::All,
+                    ..Default::default()
+                },
+            );
+
+            // Read the source file to find all pub item names — these are what
+            // callers might import individually
+            let source_content = std::fs::read_to_string(&source_abs).unwrap_or_default();
+            let pub_item_names: Vec<String> = if let Some(items) =
+                core_parse_items(ext, &source_content)
+                    .or_else(|| ext_parse_items(ext, &source_content, from))
+            {
+                items
+                    .iter()
+                    .filter(|item| item.visibility == "pub" || item.visibility == "pub(crate)")
+                    .map(|item| item.name.clone())
+                    .collect()
+            } else {
+                vec![]
+            };
+
+            // Also include the module name itself for `use crate::path::module;` imports
+            let mut search_terms: Vec<String> = pub_item_names.clone();
+            search_terms.push(source_module_leaf.to_string());
+
+            let search_refs: Vec<&str> = search_terms.iter().map(|s| s.as_str()).collect();
+
+            for file_path in &all_files {
+                let rel_path = file_path
+                    .strip_prefix(root)
+                    .unwrap_or(file_path)
+                    .to_string_lossy()
+                    .to_string();
+
+                if rel_path == from || rel_path == to {
+                    continue;
+                }
+
+                let file_ext_str = file_path.extension().and_then(|e| e.to_str()).unwrap_or("");
+                if !ext.handles_file_extension(file_ext_str) {
+                    continue;
+                }
+
+                let file_content = match std::fs::read_to_string(file_path) {
+                    Ok(c) => c,
+                    Err(_) => continue,
+                };
+
+                // Quick check: does this file mention the source module?
+                let mentions_source = search_refs.iter().any(|term| file_content.contains(term));
+                if !mentions_source {
+                    continue;
+                }
+
+                if let Some(rewrites) = ext_rewrite_caller_imports(
+                    ext,
+                    &search_refs,
+                    &source_module,
+                    &dest_module,
+                    &file_content,
+                    &rel_path,
+                ) {
+                    if !rewrites.is_empty() {
+                        imports_updated += rewrites.len();
+                        caller_rewrites.push((file_path.to_path_buf(), rewrites));
+                    }
+                }
+            }
+        }
+    }
+
+    // ── Phase 2: Update mod.rs declarations ──────────────────────────
+    let mut mod_declarations_updated = false;
+
+    // Find old parent mod.rs and remove `mod source_stem;`
+    let old_parent_mod = find_parent_mod_rs(from, root);
+    let new_parent_mod = find_parent_mod_rs(to, root);
+
+    // ── Phase 3: Apply if requested ─────────────────────────────────
+    if write {
+        // Create destination parent directory
+        if let Some(parent) = dest_abs.parent() {
+            std::fs::create_dir_all(parent).map_err(|e| {
+                crate::Error::internal_io(
+                    e.to_string(),
+                    Some(format!("create directory {}", parent.display())),
+                )
+            })?;
+        }
+
+        // Move the file
+        std::fs::rename(&source_abs, &dest_abs).map_err(|e| {
+            crate::Error::internal_io(e.to_string(), Some(format!("move {} → {}", from, to)))
+        })?;
+
+        // Update old parent mod.rs — remove mod declaration
+        if source_stem != "mod" {
+            if let Some(ref old_mod) = old_parent_mod {
+                if old_mod.exists() {
+                    if let Ok(content) = std::fs::read_to_string(old_mod) {
+                        let new_content = remove_mod_declaration(&content, source_stem);
+                        if new_content != content {
+                            let _ = std::fs::write(old_mod, &new_content);
+                            mod_declarations_updated = true;
+                        }
+                    }
+                }
+            }
+        }
+
+        // Update new parent mod.rs — add mod declaration
+        if dest_stem != "mod" {
+            if let Some(ref new_mod) = new_parent_mod {
+                if new_mod.exists() {
+                    if let Ok(content) = std::fs::read_to_string(new_mod) {
+                        let new_content = add_mod_declaration(&content, dest_stem);
+                        if new_content != content {
+                            let _ = std::fs::write(new_mod, &new_content);
+                            mod_declarations_updated = true;
+                        }
+                    }
+                } else {
+                    // Create mod.rs with just the new declaration
+                    if let Some(parent) = new_mod.parent() {
+                        let _ = std::fs::create_dir_all(parent);
+                    }
+                    let _ = std::fs::write(new_mod, format!("mod {};\n", dest_stem));
+                    mod_declarations_updated = true;
+                }
+            }
+        }
+
+        // Apply caller import rewrites
+        for (file_path, rewrites) in &caller_rewrites {
+            if let Ok(content) = std::fs::read_to_string(file_path) {
+                let mut lines: Vec<&str> = content.split('\n').collect();
+                // Apply rewrites in reverse line order to preserve line numbers
+                let mut sorted_rewrites = rewrites.clone();
+                sorted_rewrites.sort_by(|a, b| b.line.cmp(&a.line));
+                for rewrite in &sorted_rewrites {
+                    let idx = rewrite.line.saturating_sub(1);
+                    if idx < lines.len() {
+                        // Handle multi-line replacements (e.g., split grouped imports)
+                        let replacement_lines: Vec<&str> =
+                            rewrite.replacement.split('\n').collect();
+                        lines.splice(idx..=idx, replacement_lines);
+                    }
+                }
+                let new_content = lines.join("\n");
+                let _ = std::fs::write(file_path, new_content);
+            }
+        }
+    }
+
+    let caller_files: Vec<String> = caller_rewrites
+        .iter()
+        .map(|(path, _)| {
+            path.strip_prefix(root)
+                .unwrap_or(path)
+                .to_string_lossy()
+                .to_string()
+        })
+        .collect();
+
+    Ok(MoveFileResult {
+        from_file: from.to_string(),
+        to_file: to.to_string(),
+        imports_updated,
+        caller_files_modified: caller_files,
+        applied: write,
+        warnings,
+        mod_declarations_updated,
+    })
+}
+
+/// Find the parent module's mod.rs (or lib.rs) for a given file path.
+fn find_parent_mod_rs(file_path: &str, root: &Path) -> Option<PathBuf> {
+    let path = Path::new(file_path);
+    let parent = path.parent()?;
+    let mod_rs = root.join(parent).join("mod.rs");
+    if mod_rs.exists() {
+        return Some(mod_rs);
+    }
+    // Check for lib.rs in the parent (src/lib.rs)
+    let lib_rs = root.join(parent).join("lib.rs");
+    if lib_rs.exists() {
+        return Some(lib_rs);
+    }
+    None
+}
+
+/// Remove a `mod foo;` declaration from module content.
+fn remove_mod_declaration(content: &str, module_name: &str) -> String {
+    let patterns = [
+        format!("pub mod {};", module_name),
+        format!("pub(crate) mod {};", module_name),
+        format!("pub(super) mod {};", module_name),
+        format!("mod {};", module_name),
+    ];
+    content
+        .lines()
+        .filter(|line| {
+            let trimmed = line.trim();
+            !patterns.iter().any(|p| trimmed == p)
+        })
+        .collect::<Vec<_>>()
+        .join("\n")
+        + if content.ends_with('\n') { "\n" } else { "" }
+}
+
+/// Add a `mod foo;` declaration to module content (after existing mod declarations).
+fn add_mod_declaration(content: &str, module_name: &str) -> String {
+    // Check if it already exists
+    let patterns = [
+        format!("pub mod {};", module_name),
+        format!("pub(crate) mod {};", module_name),
+        format!("mod {};", module_name),
+    ];
+    if content
+        .lines()
+        .any(|line| patterns.iter().any(|p| line.trim() == p))
+    {
+        return content.to_string();
+    }
+
+    // Find the last `mod` or `pub mod` line and insert after it
+    let lines: Vec<&str> = content.lines().collect();
+    let mut insert_after = None;
+    for (i, line) in lines.iter().enumerate() {
+        let trimmed = line.trim();
+        if trimmed.starts_with("mod ")
+            || trimmed.starts_with("pub mod ")
+            || trimmed.starts_with("pub(crate) mod ")
+        {
+            insert_after = Some(i);
+        }
+    }
+
+    let new_line = format!("pub mod {};", module_name);
+    let mut result: Vec<String> = lines.iter().map(|l| l.to_string()).collect();
+    if let Some(idx) = insert_after {
+        result.insert(idx + 1, new_line);
+    } else {
+        // No existing mod declarations — prepend
+        result.insert(0, new_line);
+        result.insert(1, String::new());
+    }
+
+    let mut out = result.join("\n");
+    if content.ends_with('\n') && !out.ends_with('\n') {
+        out.push('\n');
+    }
+    out
+}
+
 // ============================================================================
 // Helpers
 // ============================================================================
+
+/// Resolve the root directory from component ID or explicit path.
+pub fn resolve_root(component_id: Option<&str>, path: Option<&str>) -> Result<PathBuf> {
+    if let Some(p) = path {
+        let pb = PathBuf::from(p);
+        if !pb.is_dir() {
+            return Err(crate::Error::validation_invalid_argument(
+                "path",
+                format!("Not a directory: {}", p),
+                None,
+                None,
+            ));
+        }
+        Ok(pb)
+    } else {
+        let comp = component::resolve(component_id)?;
+        let validated = component::validate_local_path(&comp)?;
+        Ok(validated)
+    }
+}
 
 // ============================================================================
 // Tests
